fixed tables always fetching data to fetching data if table is open and display position sets position to current place instead of animating to position
