+++ app/src/components/map/ToolbarContainer.tsx
@@ -97,7 +97,7 @@ export const ToolbarContainer = (props) => {
         <LayerPicker inputGeo={props.inputGeo} />
         <Divider />
         <SetPointOnClick map={props.map} />
-        {/* <DisplayPosition map={props.map} /> */}
+        <DisplayPosition map={props.map} />
         <MeasureTool />
         <ZoomControl mapMaxNativeZoom={props.mapMaxNativeZoom} setMapMaxNativeZoom={props.setMapMaxNativeZoom} />
         {Capacitor.getPlatform() !== 'web' ? <JumpToTrip /> : <></>}
